微信开放平台授权登陆

+++ common-utils/src/main/java/org/igetwell/common/bean/WxOpenComponentAuthorization.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class WxOpenComponentAuthorization {
     private String authorizerAppid;
     private String authorizerAccessToken;
-    private int expiresIn;
+    private long expiresIn;
     private String authorizerRefreshToken;
     private List<Integer> funcInfo;
 }
